Updated(VoxelState): Memory optimization by packing fields into single ushort and using bitwise operations to unpack them & Fixed block face rendering, again

+++ Assets/Scripts/Player.cs
@@ -43,7 +43,7 @@ public class Player : MonoBehaviour
     private float lastMoveSpeed;
     private bool jumpRequest;
 
-    public int orientation;
+    public byte orientation;
 
     [Header("Block Destroy & Placement properties")]
     public bool showHighlightBlocks = true;
